pyproject.toml
[tool.black]
line-length=79
skip-magic-trailing-comma=true

[tool.poetry]
name = "pycooldown"
version = "0"
description = "A lightning fast cooldown/ratelimit implementation."
authors = ["Circuit <circuitsacul@icloud.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/TrigonDev/pycooldown"
repository = "https://github.com/TrigonDev/pycooldown"
keywords = ["cython", "ratelimit", "cooldown"]
build = "build.py"

[tool.poetry.dependencies]
python = "^3.8"
mypy-extensions = "^0.4.3"

[tool.poetry.dev-dependencies]
nox = "^2022.1.7"
mypy = "^0.991"
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pytest-cov = "^4.0.0"
pytest-mock = "^3.7.0"
mock = "^4.0.3"

[build-system]
requires = [
    "poetry-core>=1.0.0",
    "setuptools",
    "mypy",
]
build-backend = "poetry.core.masonry.api"
